\name{plot.fRFE}
\alias{plot.fRFE}
\title{
    Plot function for object of class fRFE
}
\description{
    Plot the error rates or MSE of an object of class fRFE which is the result of the functions \code{\link{selectFunctional}}, \code{\link{selectGroup}} and \code{\link{selectLevel}}.
}
\usage{
\method{plot}{fRFE}(x, ...)
}
\arguments{
  \item{x}{
    An object of class fRFE.
}
  \item{\dots}{
    optional graphical parameters.
}
}



\author{
    Baptiste Gregorutti
}
