\name{print.fRFE}
\alias{print.fRFE}
\title{
    Print function for object of class fRFE
}
\description{
    A print function of an object of class fRFE which is the result of the functions \code{\link{selectFunctional}}, \code{\link{selectGroup}} and \code{\link{selectLevel}}.
}
\usage{
\method{print}{fRFE}(x, ...)
}
\arguments{
  \item{x}{
    An object of class fRFE.
}
  \item{\dots}{
    optional parameters to be passed to the function \sQuote{summary}.
}
}

\author{
	Baptiste Gregorutti
}
