\name{RFgroove-package}
\alias{RFgroove-package}
\alias{RFgroove}
\docType{package}
\title{
Importance Measure and Selection for Groups of Variables with Random Forests
}
\description{
	Variable selection tools for groups of variables and multivariate functional data based on a new grouped variable importance with Random Forests.
}
\details{
\tabular{ll}{
Package: \tab RFgroove\cr
Type: \tab Package\cr
Version: \tab 1.1\cr
Date: \tab 2016-03-16\cr
License: \tab GPL (>=2.0)\cr
}

}
\author{
	Baptiste Gregorutti
	Maintainer: Baptiste Gregorutti <baptiste.gregorutti@safety-line.fr>
}
\references{
    Gregorutti, B., Michel, B. and Saint Pierre, P. (2015). Grouped variable importance with random forests and application to multiple functional data analysis, Computational Statistics and Data Analysis 90, 15-35.
}
