\name{summary.fRFE}
\alias{summary.fRFE}
\title{
    Summary function for object of class fRFE
}
\description{
    A summary function of an object of class fRFE which is the result of the functions \code{\link{selectFunctional}}, \code{\link{selectGroup}} and \code{\link{selectLevel}}.
}
\usage{
\method{summary}{fRFE}(object, ...)
}
\arguments{
  \item{object}{
    An object of class fRFE.
}
  \item{\dots}{
    further arguments passed to or from other methods.
}
}
\author{
    Baptiste Gregorutti
}
