\name{toyClassif}
\alias{toyClassif}
\docType{data}
\title{
	A toy data of classification
}
\description{
	A toy data of classification drawn from a mixture Gaussian distribution. Only the three first variables are discriminant.
}
\usage{data(toyClassif)}
\format{
  A list of two components:
	\itemize{
		\item X: the design matrix of size \eqn{n \times p} drawn from a mixture Gaussian distribution ;
		\item Y: the class labels.
	}
}
