\name{toyReg}
\alias{toyReg}
\docType{data}
\title{
	A toy data of regression
}
\description{
	A toy data of regression drawn from a multivariate Gaussian distribution. Only the three first variables are discriminant.
}
\usage{data(toyReg)}
\format{
  A list of two components:
	\itemize{
		\item X: the design matrix of size \eqn{n \times p} drawn from a multivariate Gaussian distribution ;
		\item Y: the outcome drawn from the Gaussian distribution.
	}
}
